This is the results of my JDL test as an example.
1. I created the `jhipster-jdl-2.jh` file using https://start.jhipster.tech/jdl-studio/.
1. Then I used the command `jhipster import-jdl ~/Downloads/jhipster-jdl-2.jh` in the root folder to generate all the files.
1. Lastly, I did a SQL table copy into the generated SKU table.

Once it was running, I added the contents of the SKU table into Liquibase so the database is created with the correct data automatically.

This code is just to get a feel for what can be done with JDL and code generation. It is nowhere near powerful enough to to be used for the actual project results.